requirements.txt
pyaml
boto3